# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_thienel_go-backend-template_internal_interface_api_dto.AssignerProfileUUID:
    properties:
      profile_uuid:
        type: string
    type: object
...
  github_com_thienel_go-backend-template_internal_interface_api_dto.DEPConfig:
    properties:
      base_url:
        type: string
    type: object
  github_com_thienel_go-backend-template_internal_interface_api_dto.DEPDevicesRequest:
    properties:
      devices:
        example:
        - G0NXR2GUKPFQ
        items:
          type: string
        type: array
    type: object
...
  github_com_thienel_go-backend-template_pkg_response.APIResponse-any:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.Error'
      is_success:
        type: boolean
      message:
        type: string
    type: object
...
  ? github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_AssignerProfileUUID
  : properties:
      data:
        $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.AssignerProfileUUID'
      error:
        $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.Error'
      is_success:
        type: boolean
      message:
        type: string
    type: object
  ? github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_BypassCodeResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.BypassCodeResponse'
      error:
        $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.Error'
      is_success:
        type: boolean
      message:
        type: string
    type: object
...
  ? github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_DEPConfig
  : properties:
      data:
        $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.DEPConfig'
      error:
        $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.Error'
      is_success:
        type: boolean
      message:
        type: string
    type: object
  ? github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_DEPNamesQueryResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.DEPNamesQueryResponse'
      error:
        $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.Error'
      is_success:
        type: boolean
      message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Server for THD iOS MDM Portal.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: THD iOS MDM Portal API
  version: "1.0"
paths:
...
  /api/v1/dep/assigner/{name}:
    get:
      description: Return the assigner profile UUID for the given DEP name.
      parameters:
      - description: Name of DEP server instance
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_AssignerProfileUUID'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Return the assigner profile UUID
      tags:
      - DEP Setup
    put:
      description: Assign a profile UUID for assignment for the given DEP name.
      parameters:
      - description: Name of DEP server instance
        in: path
        name: name
        required: true
        type: string
      - description: Profile UUID to assign
        in: query
        name: profile_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_AssignerProfileUUID'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Assign a profile UUID for assignment
      tags:
      - DEP Setup
  /api/v1/dep/bypasscode:
    get:
      description: Generates (or decodes) an Activation Lock Bypass Code and returns
        different forms of it.
      parameters:
      - description: Hex-encoded raw form of bypass code
        in: query
        name: code
        type: string
      - description: Dash-separated human readable form
        in: query
        name: raw
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_BypassCodeResponse'
      security:
      - BearerAuth: []
      summary: Generates or decodes an Activation Lock Bypass Code
      tags:
      - DEP Setup
  /api/v1/dep/config/{name}:
    get:
      description: Return the config for the given DEP name.
      parameters:
      - description: Name of DEP server instance
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_DEPConfig'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Return the config for the given DEP name
      tags:
      - DEP Setup
    put:
      consumes:
      - application/json
      description: Set the config for the given DEP name.
      parameters:
      - description: Name of DEP server instance
        in: path
        name: name
        required: true
        type: string
      - description: Config details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.DEPConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_DEPConfig'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Set the config for the given DEP name
      tags:
      - DEP Setup
  /api/v1/dep/dep_names:
    get:
      description: Query DEP names with optional filters and pagination.
      parameters:
      - collectionFormat: csv
        description: Filter by DEP names
        in: query
        items:
          type: string
        name: dep_name
        type: array
      - default: 100
        description: Limits number of results
        in: query
        name: limit
        type: integer
      - description: Offset results
        in: query
        name: offset
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_DEPNamesQueryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Query DEP names
      tags:
      - DEP Setup
  /api/v1/dep/maidjwt/{name}:
    get:
      description: Generate Managed Apple ID Managed Access JWT.
      parameters:
      - description: Name of DEP server instance
        in: path
        name: name
        required: true
        type: string
      - description: MDM server UUID
        in: query
        name: server_uuid
        type: string
      produces:
      - application/jwt
      responses:
        "200":
          description: JWT
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Generate Managed Apple ID Managed Access JWT
      tags:
      - DEP Setup
  /api/v1/dep/profiles:
    get:
      description: Fetch all defined DEP profiles from NanoDEP
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: List Defined DEP profiles
      tags:
      - DEP Setup
  /api/v1/dep/proxy/{name}/account:
    get:
      description: Fetch Apple DEP account info via proxy
      parameters:
      - description: DEP name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get DEP account
      tags:
      - DEP Setup
  /api/v1/dep/proxy/{name}/devices:
    post:
      consumes:
      - application/json
      description: Fetch device list or details from Apple via proxy. If devices array
        is provided in body, fetches details for those serial numbers.
      parameters:
      - description: DEP name
        in: path
        name: name
        required: true
        type: string
      - description: Devices list
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.DEPDevicesRequest'
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get DEP devices
      tags:
      - DEP Setup
  /api/v1/dep/proxy/{name}/devices/disown:
    post:
      consumes:
      - application/json
      description: Remove a device from DEP management via proxy
      parameters:
      - description: DEP name
        in: path
        name: name
        required: true
        type: string
      - description: Devices list to disown
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.DEPDevicesRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Disown DEP device
      tags:
      - DEP Setup
  /api/v1/dep/proxy/{name}/devices/sync:
    post:
      description: Initiate a sync with Apple DEP servers to fetch new devices via
        proxy
      parameters:
      - description: 'DEP name (default: ''default'')'
        in: path
        name: name
        type: string
      - description: Sync cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Sync DEP devices
      tags:
      - DEP Setup
  /api/v1/dep/proxy/{name}/profile:
    get:
      description: Fetch details of a defined DEP profile via proxy
      parameters:
      - description: 'DEP name (default: ''default'')'
        in: path
        name: name
        type: string
      - description: Profile UUID
        in: query
        name: profile_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Get DEP profile
      tags:
      - DEP Setup
    post:
      consumes:
      - application/json
      description: Create or update a DEP assignment profile via proxy
      parameters:
      - description: 'DEP name (default: ''default'')'
        in: path
        name: name
        type: string
      - description: Profile details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.DEPProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_DEPProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Define DEP profile
      tags:
      - DEP Setup
  /api/v1/dep/token/{name}:
    get:
      description: Get PEM certificate for ABM for a specific name
      parameters:
      - description: Token name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Get DEP token pem
      tags:
      - DEP Setup
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...